setting refactor
