+++ src/plugins/open-ai-embeddings/server/initialize.js
@@ -23,6 +23,7 @@ class PluginManager {
       return { pinecone: this.pinecone, index: this.index, pineconeStore: this.pineconeStore };
     } catch (error) {
       console.error(`Failed to initialize Pinecone: ${error}`);
+      throw new Error(`Failed to initialize Pinecone: ${error}`);
     }
   }
 
@@ -40,6 +41,7 @@ class PluginManager {
       return this.embeddings;
     } catch (error) {
       console.error(`Failed to initialize Embeddings: ${error}`);
+      throw new Error(`Failed to initialize Embeddings: ${error}`);
     }
   }
 
@@ -52,29 +54,33 @@ class PluginManager {
       return this.pineconeStore;
     } catch (error) {
       console.error(`Failed to initialize Pinecone Store: ${error}`);
+      throw new Error(`Failed to initialize Pinecone Store: ${error}`);
     }
   }
 
-  async initializeModel(openAIApiKey) {
+  async initializeModel(openAIApiKey, modelName) {
     if (this.model) return this.model;
-    // TODO: SET MODEL NAME IN SETTINGS
     try {
       const model = new OpenAI({
         openAIApiKey: openAIApiKey,
-        modelName: "gpt-3.5-turbo",
+        modelName: modelName,
       });
       this.model = model;
       return this.model;
     } catch (error) {
       console.error(`Failed to initialize Model: ${error}`);
+      throw new Error(`Failed to initialize Model: ${error}`);
     }
   }
 
   async initialize(settings) {
-    await this.initializePinecone(settings.pineConeApiEnv, settings.pineConeApiKey, "strapi-plugin");
-    await this.initializeEmbeddings(settings.apiKey);
+    await this.initializePinecone(settings.pineconeApiEnv, settings.pineconeApiKey, settings.pineconeIndex);
+    await this.initializeEmbeddings(settings.openAiApiKey);
     await this.initializePineconeStore();
-    await this.initializeModel(settings.apiKey);
+    await this.initializeModel(settings.openAiApiKey, settings.openAiModelName);
+  }
+
+  async getSettings() {
     return { pinecone: this.pinecone, pineconeIndex: this.index, pineconeStore: this.pineconeStore, model: this.model };
   }
 }
